+++ src/main/java/appland/upload/UploadResponse.java
@@ -1,0 +1,10 @@
+package appland.upload;
+
+import com.google.gson.annotations.SerializedName;
+
+class UploadResponse {
+    @SerializedName("id")
+    Integer id;
+    @SerializedName("token")
+    String token;
+}